Cập nhật danh mục loại mặt hàng

+++ soure_code/StorageManager/src/main/java/com/hinet/storage/dao/ProductCategoryDAO.java
@@ -21,7 +21,7 @@ public interface ProductCategoryDAO {
 
     boolean updateProductCategory(ProductCategory action);
 
-    boolean deleteProductCategory(long id);
+    boolean deleteProductCategory(int id);
 
-    ProductCategory getProductCategoryById(long id);
+    ProductCategory getProductCategoryById(int id);
 }
